Add files via upload
This is all the code for Binary Search Tree. At the moment it works completely. It uses InputData.txt for the data and uses two lines for the data. We can change this so that it takes more than two lines for one BSTNode. It has one tree and can print it 2D and in all the traversal methods. It can delete and modify values. These deletes and modifications are saved to the InputData.txt file.